Geolocation: refactor and separate geolocate/1 from reverse_geocode/1
Refactor the fetch_location/1 function in the Geolocation context by
splitting it into two distinct functions, geolocate/1 and
reverse_geocode/1.

This approach helps separating the concerns of the two computations,
also mapping to the already existing distinction between geolocation
and geocoding modules:
- Geolocation modules accept a Device and compute a Position, i.e.
timestamped GPS coordinates.
- Reverse geocoding modules accept some Coordinates and compute a
textual Location, i.e. a place address.

A Location type is also defined, which currently maps to String.t() but
could be easily refactored in the future as a struct that holds
individual address parts.

Signed-off-by: Davide Briani <davide.briani@secomind.com>